Start encapsulating `BaseExpression` (https://redirect.github.com/duckdb/duckdb/pull/15360)
This PR adds new getters and setters to the `BaseExpression` class and
updates core and the json and parquet extensions to use these new
accessors instead of directly accessing the class members.

The initial plan was to make the underlying class members protected (or
private), but as that would break almost all out-of-tree extensions I
figured it would be better to try to do this in small steps and try to
get this merged first so that I can update the extension incrementally
later without having to patch everything immediately.

Upstream-commit: https://github.com/duckdb/duckdb/commit/48c9f5e79676d93f0e2bf64197521cd8bc6b2e89
--- 

I also think it would be a good idea to add some helpers to the
BaseExpression class to make it slightly easier to read code that does a
lot of matching/switching on the expression type/class. Something like
```
bool BaseExpression::IsType(ExpressionType type) const;

bool BaseExpression::IsOneOfType(ExpressionType type1, ... types) const
```

But that's for another time.