feat: refactor getting tie_break_offset in SelectBestMatch (https://redirect.github.com/duckdb/duckdb/pull/15235)
We have a case where we need access to the tie break offset of different
secret storages when we do `LookupSecrets` in one secret storage (the
one on the server side). This refactor adds "tie break offset" of a
secret storage to the parameter list of `SelectBestMatch` method so that
this function can be more flexibly used without needing to rely on the
class it belongs to (GetTiebreakOffset()). As a result, I'm also
changing this function to `static` (still `protected`) in SecretStorage
to make it clear that it's not tied to any class instance and adding
`tie_break_offset` to be a const parameter when instantiating a
SecretStorage class to ensure that this value doesn't get modified after
initialization.

Upstream-commit: https://github.com/duckdb/duckdb/commit/ff92dbe95021d1cd7331406662c63b07c420868f