Type mismatch set operation  (https://redirect.github.com/duckdb/duckdb/pull/15422)
fixes https://github.com/duckdblabs/duckdb-internal/issues/3638

Classic case of "3 hours to find the bug, 15 minutes to fix it"

If we introduce a delim get using a cross join under a set operation, we
need to also add a projection above the cross join. Otherwise, if the
children of the cross join get swapped, the columns going into the set
operation are out of order, hence why this bug occurred.

Upstream-commit: https://github.com/duckdb/duckdb/commit/996e8561f5d779c9fe70d71ebb1520dd815321ee