Retain join partition order (https://redirect.github.com/duckdb/duckdb/pull/15460)
Follow-up of https://redirect.github.com/duckdb/duckdb/pull/15389, which improved partition selection by choosing
small partitions first.

An oversight of that PR was that completely relying on sizes to select
partitions will shuffle the partition order even if there are tiny
differences which we wouldn't really care about. This PR rounds the
partition sizes down to a multiple of the smallest partition size before
sorting and using a stable sort. This way, tiny differences in partition
sizes don't affect the sort order, and we will retain the original order
as much as possible.

I've added a benchmark file, but was not able to see a significant
difference with and without the change. However, counted I/O calls and
saw that read/writes were reduced by around 5%, so this should help more
on machines with slower disks.

Upstream-commit: https://github.com/duckdb/duckdb/commit/e7765f25f1345597e4d0509a0f0248680f68da20