Allow a variable type `rowid` pseudocolumn in tables (https://redirect.github.com/duckdb/duckdb/pull/14674)
This PR introduces support for configuring the `rowid` pseudo-column to
use a type other than `BIGINT`. In particular, it enables `rowid` to be
defined as either a `BLOB` or `VARCHAR`. I think this is useful when
working with external data sources, and especially Apache Arrow provided
data sources. This PR is one step on the road for enabling
`DELETE`/`UPDATE`/`INSERT` to tables via Apache Arrow Flight but could
be used by other data sources.

### Changes

To support this, I’ve updated the `TableCatalogEntry` to specify the
type of `rowid` as a `LogicalType`, so this change can be used by other
catalog implementations.

@pdet, In the Arrow integration, `ArrowTableFunction::ArrowToDuckDB` now
accepts an optional `rowid_column_index` parameter to indicate the
position of the `rowid` column within the Arrow table. This change
ensures that `rowid` can appear at any position within an Arrow array.

To fully support this functionality, I also adjusted
`BindContext::AddBaseTable`, and the `LogicalGet`, and `Binding`
classes, ensuring that the `rowid` column type is dynamic but defaults
to `LogicalTypeId::ROW_TYPE` if unspecified.

### Discussion

I realize this PR is a big bigger than the size I typically send, and
will require more energy to review. I tried to make it as small as
possible, but if there is a better way to accomplish what I'm trying to
do I'm happy to rework the approach. I’d also be happy to discuss any
details or adjustments to make this approach work smoothly for others.
Thanks for reviewing this.

Upstream-commit: https://github.com/duckdb/duckdb/commit/13ba13c121acfb3f4c48c16337297ac705779c19