[Julia]: Auto-generate api.jl  (requires duckdb v1.2?) (https://redirect.github.com/duckdb/duckdb/pull/15474)
This PR adds a python script (`generate_c_api_julia.py`) to
automatically generate function definitions for the DuckDB C-API in
`api.jl`. It also generates documentation and deprecation warning (via
`Base.depwarn`) if this information is given.

It fixes various types, where `api.jl` had mismatches between the types
that where defined in `duckdb.h` and uncovered to usage of some
deprecated functions. Also some functions to convert between types have
been added to `ctypes.jl`.

It uses a similiar logic that is used to generate `duckdb.h` and uses
the definitions
from`src/include/duckdb/main/capi/header_generation/apis`. These are
created for `v1.2`, so this changes most likely require this version as
well.

I hope this can be used to keep the Julia Pkg more easily up-to-date
with DuckDB :)


_Note_: For easy review, the original order of the functions is kept.
This can be changed in a future PR to be consistent to `duckdb.h`

Upstream-commit: https://github.com/duckdb/duckdb/commit/b4e3d0ed54f1ccd0074e6ae4746a3aa98c504f55