[Julia]: Auto-generate api.jl with new order (https://redirect.github.com/duckdb/duckdb/pull/15580)
This is a follow up PR to https://redirect.github.com/duckdb/duckdb/pull/15474 that changes the order in which the
functions in `api.jl` are defined. They now have the same order as in
`duckdb.h` and with proper grouping.

To switch back to the original order the `--use-original-order` flag in
the `update_api.sh` script can be set.

In the previous PR the original order was intentionally kept to more
easily spot discrepancies between the hand-coded and automatically
generated wrapper and to keep the git diff to a minimum for review.

Upstream-commit: https://github.com/duckdb/duckdb/commit/b9e368e888fed036f598acf4994bd30fbfe97472