In PhysicalInsert call FinalFlush before merging row groups into local storage (https://redirect.github.com/duckdb/duckdb/pull/15583)
This fixes an issue where sporadically we could run into a nullpointer
exception when flushing partial blocks later on.

Essentially what could happen is that we would have blocks that were
added to the partial block manager in the transaction-local storage get
appended to later on again, leading to the same segment being registered
with multiple partial blocks. Flushing the partial block manager before
merging the segments resolves this issue as it ensures the segments are
converted into disk-backed persistent blocks, which prevents them from
being appended to again later on.

Upstream-commit: https://github.com/duckdb/duckdb/commit/41d13ad1612b3e91ff7a9a5c17edd6fdf51073f1