InternalException should only invalidate database when encountered during execution (https://redirect.github.com/duckdb/duckdb/pull/15588)
Internal exceptions signal failures of internal assertions in DuckDB -
as such we force the database to be shut down, as they might leave
DuckDB in an unknown state.

Most of the internal exceptions we encounter, however, happen during the
bind/optimization phases - and these do not have any danger of leaving
the database in an unknown state. This PR makes it so that only internal
exceptions encountered during the execution or commit/rollback phases
invalidate the database.

Upstream-commit: https://github.com/duckdb/duckdb/commit/3c0db29b2bf182105c1537ddf2fa8b12d186941d