More bugfixes (https://redirect.github.com/duckdb/duckdb/pull/15605)
Fixes:
1. https://redirect.github.com/duckdb/duckdb/pull/15432
2. https://redirect.github.com/duckdb/duckdb/pull/15504
3. [This
comment](https://github.com/duckdb/duckdb/commit/04d20edf72c97ba79b374f2a6a8cbdf35cec1231#r150914884)

The issue for (1.) was that macros' dummy bindings still had the type ID
`SQLNULL` when they should have had the type ID `UNKNOWN`. I've
corrected this, which fixed the issue. This also uncovered some other
issues, which could be resolved by throwing
`ParameterNotResolvedException` in some places.

The issue for (2.) was that Parquet files use 3 schema elements for
lists, while we only use 2. The `field_id` of the parent was set
correctly, but this schema element is empty, so we skip it when
converting to DuckDB lists. The child schema element we read does not
have a `field_id,` so it was assumed not to be present in the Parquet
file. I've fixed this by optionally checking the `field_id` of the
parent schema element.

Finally, for (3.), I added an optimization that was never triggered. We
now properly compare pointers and trigger the optimization, which allows
us to skip `memcpy` when `ReAllocate`ing memory that was allocated by
the `ArenaAllocator`.

EDIT: I've also added a line that resets the `dict_state` in
`AggregateHashTable` when we call `Abandon()` in an attempt to fix
https://github.com/duckdblabs/duckdb-internal/issues/3826 (which I
cannot reproduce locally).

Upstream-commit: https://github.com/duckdb/duckdb/commit/ce33de966a521d1a6e86ec9579e133ff2b2534f4