fix creating VARINT logical type in C API (https://redirect.github.com/duckdb/duckdb/pull/15670)
See https://github.com/duckdb/duckdb/issues/15669.

There was a missing case for VARINT in `ConvertCTypeToCPP`, which meant
attempting to create a VARINT type through the C API (using
`duckdb_create_logical_type(DUCKDB_TYPE_VARINT)`) instead returned a
`duckdb_logical_type` with the type id `DUCKDB_TYPE_VARINT`.

Among other things, this meant attempting to create a data chunk with a
VARINT column didn't work. That's what I was trying to do when I found
this. I included two unit tests for this case, one for VARINT (which was
previously broken) and one for BLOB (which was not).

Upstream-commit: https://github.com/duckdb/duckdb/commit/db06cdd9ddb420a7f04248e7a62ea1785868d97c