Allow switching to a different catalog from a detached catalog (https://redirect.github.com/duckdb/duckdb/pull/15624)
In a multi-connection setting, if con1's default database is detached by
another connection, switching to a different catalog using `USE
other_db` in con1 throws a binder error. The reason is that DuckDB calls
`GetCatalog` for the detached default database to search if the
`other_db` matches one of the schemas there. A workaround was to use a
fully qualified name: `SET search_path='other_db.main'`.

This PR allows users to get out of the detached catalog using the `USE`
statement.

Upstream-commit: https://github.com/duckdb/duckdb/commit/4e3920c3ec63284d11411d1ac6b28af5ceb784a0