Clean up temporary test directory in `run_tests_one_by_one.py` even if test segfaults (https://redirect.github.com/duckdb/duckdb/pull/15688)
The cleanup is built into our `unittest` executable, but the cleanup
doesn't happen if a test segfaults. If the test that segfaults uses a
lot of space, subsequent tests may run out of disk space (GH actions
don't have that much disk space). This PR makes sure the test dir is
cleaned up in such cases.

Upstream-commit: https://github.com/duckdb/duckdb/commit/2e533ec9dfaa82baac24ae1104e501af723a565a