Add late_materialization_max_rows setting that allows you to configure the threshold at which late materialization is triggered (https://redirect.github.com/duckdb/duckdb/pull/15697)
Follow-up from https://github.com/duckdb/duckdb/pull/15692

This PR adds the `late_materialization_max_rows` setting that allows you
to configure the threshold at which late materialization is triggered.
The default value is `50`.

Example usage:

```sql
SET late_materialization_max_rows=1000;
explain SELECT * FROM lineitem ORDER BY l_orderkey DESC LIMIT 1000;
```

The exact best setting is hard to determine - essentially the row-id
pushdown has two components to it (1) the OR filter push-down, which is
done for up to `dynamic_or_filter_threshold` rows (defaults to 50) and
the min-max filter pushdown.

The row-id rewrite generally always provides performance improvements
for up to `dynamic_or_filter_threshold` which is why we select that as a
default. Beyond that, it depends on the locality of the rows. If the
min-max filter on row-ids is selective (i.e. the rows we select are
close together physically in the table) the row-id rewrite is effective.
If the rows are spread out, the rewrite can worsen performance.

CC @abramk

Upstream-commit: https://github.com/duckdb/duckdb/commit/e0d79a7eb019477b12976b33352a2370d48d2cad