Preserve stack trace information when re-throwing (https://redirect.github.com/duckdb/duckdb/pull/15709)
When executing a query like:
```cpp
con.Query("CALL throw_exception()");
```
the resulting error message includes a stack trace because
`throw_exception()` raises an internal exception.

However, the stack trace will NOT point to the table function as the
origin of the error. This happens because the exception will be
re-thrown in
https://github.com/motherduckdb/public-duckdb/blob/7ab5640bb3b03c777bc2c69a8943a37f773f4cb2/src/include/duckdb/execution/task_error_manager.hpp#L44,
thereby overwriting the original stack trace.

This PR preserves the original stack trace in case `ErrorData` is
copied.

Upstream-commit: https://github.com/duckdb/duckdb/commit/4a08b949d83e24cd02514a4dffd8def443534eb9