Make entries field non-nullable for Arrow map type (https://redirect.github.com/duckdb/duckdb/pull/15733)
`SetArrowMapFormat` in the `ArrowConverter` would produce a schema where
the nullability of the `entries` field is set to Nullable. This is
actually not correct.

Without this fix the arrow produced by DuckDB causes problems when I try
to feed it to the Delta Kernel, which uses arrow rust to decode our
schema and then complains about this nullability being incorrect.

Upstream-commit: https://github.com/duckdb/duckdb/commit/96d58253c4e116f2d9c6d3f9502eb18737ce040c