Benchmark runner: catch and log errors + add support for `retry load N` syntax (https://redirect.github.com/duckdb/duckdb/pull/16054)
Instead of terminating with an uncaught exception the benchmark runner
now logs that there was an error.

In addition, we add a new syntax `retry load N` which can be used to
retry the load phase up to `N` times in case of errors. We use this in
the h2oai benchmark (which downloads CSV files from a remote source and
sporadically runs into failures on our CI).

Upstream-commit: https://github.com/duckdb/duckdb/commit/c5b4e6338de9e5b4896a5379232f60b5dced72cb