Remove extension definitions to prevent re-compilation of the entire system on commit (https://redirect.github.com/duckdb/duckdb/pull/15955)
The extensions add an `EXT_VERSION_[XXX]` definition (e.g.
`EXT_VERSION_PARQUET`) - which defaults to the current git hash. These
definitions are not removed afterwards, leading to subsequent
compilation units to all have this definition, and stacking them up as
extensions get built, e.g.:

`autocomplete` -> `-DEXT_VERSION_AUTOCOMPLETE=\"5001ce5bc7\" `
`icu` -> `-DEXT_VERSION_AUTOCOMPLETE=\"5001ce5bc7\"
-DEXT_VERSION_ICU=\"5001ce5bc7\"`

etc...

This is not really a problem, but these definitions change when a commit
is made, leading to unnecessary recompilation. As these definitions
pollute also the compilation of the main source tree this leads to
recompiling of the entire system upon commit. By removing the
definitions we make it so that we "only" recompile the extensions on
commit instead.

Upstream-commit: https://github.com/duckdb/duckdb/commit/3f40f2188f3214c92e34964bb3edb3f9f7515324