Parquet Reader: Split `CreateReader` into two separate stages - `ParseSchema` and `CreateReader` (https://redirect.github.com/duckdb/duckdb/pull/16161)
The current Parquet reader uses the ColumnReader for two purposes:

* Parsing schema metadata (getting the types/names and reading stats)
* Reading the data

This PR splits up these two cases. The `ParquetColumnSchema` class is
added which fulfills the first purpose. The new `ParseSchema` method
looks at a Parquet schema and converts that into a
`ParquetColumnSchema`. This is used in the `ParquetReader` to obtain the
result types of the file.

The column readers are now created using the `CreateReader` method
*based on* the `ParquetColumnSchema` objects. As a result, creating the
column readers is significantly simpler (as the Parquet metadata has
already been parsed at this stage). We can also avoid creating column
readers for columns we are not going to read - which can lead to
significant performance improvements when we are only reading a small
subset of the columns when reading small-ish files.

Upstream-commit: https://github.com/duckdb/duckdb/commit/05e95a9949ad615d54b6875c023d8f6f7b761d5b