Parquet: Add dedicated Select method that can be used to push selection vectors into the read (https://redirect.github.com/duckdb/duckdb/pull/16174)
This effectively restores a previous optimization where we would skip
reading elements if they were previously filtered out. For now we only
enable this for strings - that has by far the highest performance
benefits as we can skip UTF8 validation for any strings that we don't
need to read.

For simple types like integers this optimization is not so
straightforwardly useful - as we effectively replace a `memcpy` with a
branchy lookup. I haven't run any benchmarks on this yet but I suspect
that the usefulness of this optimization depends on selectivity - i.e.
it might perform better when the selectivity is <10% (or some other to
be determined threshold). I will leave that for a future PR.

Upstream-commit: https://github.com/duckdb/duckdb/commit/4c77e9cf2143c62b84ea1bdd817f9628ac72d724