Parquet reader: batch check if buffer is available in RLEBpDecoder (https://redirect.github.com/duckdb/duckdb/pull/16185)
When reading long chains of `RleBpDecoder` data - `NextCounts` can be
called many times. Instead of checking if there is enough buffer
available for every byte we can batch check if we are certain we have
enough bytes available. The varint that we are reading is limited to max
5 bytes (as more does not fit into a `uint32_t`) so we can just call
`check_available` once. This improves performance by ~10% when reading
data with many small batches.

Upstream-commit: https://github.com/duckdb/duckdb/commit/1584c228c8a3ff716b37964f2d1577dd68c24b63