Concurrency groups for R and Wasm (https://redirect.github.com/duckdb/duckdb/pull/16201)
The latest nightly-build for `R` [was
skipped](https://github.com/duckdb/duckdb/actions/runs/13275346242/job/37063805368)
because `higher priority waiting request for 'InvokeCI' exists`. It
happened because it was in common concurrency group, but it needs its
own.

This PR adds a concurrency groups for `R` and `Wasm` workflows (in
`wasm` it was missing).

Upstream-commit: https://github.com/duckdb/duckdb/commit/f6ae05d0a23cae549c6f612026eda27130fe1600