Arena allocator for `MinMaxN` and skip `NULL`s when creating enum (https://redirect.github.com/duckdb/duckdb/pull/16246)
Instead of allocating a `std::vector` per group, we now use arena
allocations to allocate fixed-size arrays. This greatly reduces
contention on the system allocator when there are many threads/groups.
It also fixes the scaling issues encountered here:
https://duckdb.org/2024/10/25/topn.html

I also removed an error thrown when a `NULL` was encountered when
creating an `ENUM`. This was unnecessary, as we could just skip.

Upstream-commit: https://github.com/duckdb/duckdb/commit/53edbbc9b1843b67b8e1b28248fea815ff1fcc1b