Some more Parquet writer performance improvements (https://redirect.github.com/duckdb/duckdb/pull/16287)
Follow-up of https://redirect.github.com/duckdb/duckdb/pull/16243. Scraping the bottom of the barrel here, as the
previous PR got many of the biggest performance gains already.

This PR adds some more fast paths for when there are no `NULL`s, and
implements a branchless hash function for `string_t`'s that are inlined.
This required some extra care to make sure that the hash function
returns the same value whether the string is inlined or not.

Overall, the changes reduce the time it takes to write TPC-H SF10
`lineitem` to Parquet from ~2.6s to ~2.4s (with the default
`PARQUET_VERSION V1`, ~2.5s to ~2.3s with `V2`).

Upstream-commit: https://github.com/duckdb/duckdb/commit/219bafadb734fd3ebd0a383d6ff774bf5e1aa551