[Dev] Improve/Add handling of escapes in VARCHAR -> list/struct/map and align behavior (https://redirect.github.com/duckdb/duckdb/pull/15944)
This PR fixes
<https://github.com/duckdblabs/duckdb-internal/issues/3945>

The following behavior applies to casting from `VARCHAR` ->
`LIST`|`STRUCT`|`MAP`

- Backslashes outside of quotes are treated as literals, unless they are
followed by a single/double quote, in which case they act as an escape
character and are removed from the result.
- Backslashes inside of quotes are always treated as escape characters,
to use a literal backslash inside quotes, the backslash has to be
escaped (`\\`), unescaped backslashes are removed from the result.

Quotes that are used to start and end a quoted value are removed from
the result.
`[""""]` -> `[]`
`['''']` -> `[]`
`['"']` -> `["]`
`["'"]` -> `[']`

Whenever your string contains a character that is considered special by
the casting logic, it should be wrapped in quotes.
To preserve leading or trailing spaces, the value should also be wrapped
in quotes.
The list of characters that are *always*[1] special: `(whitespace)`,
`{`, `}`, `(`, `)`, `[`, `]`, `"`, `'`, `\`
Other characters that are special in certain contexts:
- STRUCT key: `:`
- MAP key: `=`
- LIST, MAP, STRUCT value: `,`

[1] STRUCT key permits the use of `{`, `}`, `(`, `)`, `[`, `]`

------------------------------------------------------------

This PR also adds support for casting from `VARCHAR` -> unnamed `STRUCT`
In the form of `( <value>, <value2>, ..., <value n> )`

Upstream-commit: https://github.com/duckdb/duckdb/commit/1e8caf82845b204f976d238d9356948d8b373538